refactor-176: update blocks component name to use new
