+++ frontend/src/pages/Blocks.tsx
@@ -1,25 +1,25 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-
-import { SortingState } from '@tanstack/react-table';
-
-import { useBlocks, IUseBlocks } from 'src/hooks';
 import {
   BlockTable,
   GradientHeading,
   PageWrapper,
   PageHead,
 } from 'src/components';
-
-import { useAppSelector } from 'src/store';
-import { Block } from 'src/api';
-
-const DEFAULT_BLOCKS_COUNT_TO_FETCH = 10;
-
-const initialParam: IUseBlocks = {
-  orderByHeight: 'desc',
-  numToShow: DEFAULT_BLOCKS_COUNT_TO_FETCH,
-};
+import {
+  useAppDispatch,
+  useAppSelector,
+  Loading,
+  getBlocks,
+  getBlockLoadingStatus,
+  fetchBlocks,
+  getTotalBlocks,
+  getBlocksTableOptions,
+  setPagination,
+} from 'src/store';
+import { SortingState } from '@tanstack/react-table';
+import { DEFAULT_PAGINATION } from 'src/api';
+// import { useAppRefresh } from 'src/hooks';
 
 const initialSorting: SortingState = [
   {
@@ -29,74 +29,102 @@ const initialSorting: SortingState = [
 ];
 
 export const Blocks: React.FC = () => {
-  const [sort, setSort] = useState<SortingState>(initialSorting);
-  const [params, setParams] = useState<IUseBlocks>(initialParam);
-  const [shouldRefetchBlocks, setShouldRefetchBlocks] = useState(false);
-
   const { refreshTimer } = useAppSelector(state => state.app);
 
+  // const { setTimer } = useAppRefresh();
+
   const { t } = useTranslation();
-  const {
-    data,
-    isLoading,
-    isFetchingNextPage,
-    fetchNextPage,
-    refetch: refetchBlocks,
-  } = useBlocks(params);
-
-  const blocks = useMemo(() => {
-    return data?.pages.reduce(
-      (accum, page) => [...accum, ...page.blocks],
-      [] as Block[],
-    );
-  }, [data]);
-
-  const total = useMemo(() => {
-    return data && data.pages.length > 0 ? data.pages[0].total : undefined;
-  }, [data]);
-
-  useEffect(() => {
-    if (refreshTimer === 0) {
-      refetchBlocks({ cancelRefetch: true });
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [refreshTimer]);
-
-  useEffect(() => {
-    if (sort.length === 0) return;
-    const { id, desc } = sort[0];
-    if (id === 'height') {
-      setParams(prev => ({
-        ...prev,
-        orderByHeight: desc ? 'desc' : 'asc',
-      }));
-      setShouldRefetchBlocks(prev => !prev);
-    }
-  }, [sort]);
-
-  useEffect(() => {
-    refetchBlocks();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [shouldRefetchBlocks]);
 
   const pageTitle = `${t('blocks')}`;
 
+  // TODO: potentially create useActions hook?
+  const dispatch = useAppDispatch();
+
+  const blocks = useAppSelector(getBlocks);
+  const totalBlocks = useAppSelector(getTotalBlocks);
+  const blockLoadingStatus = useAppSelector(getBlockLoadingStatus);
+  const blocksTableOptions = useAppSelector(getBlocksTableOptions);
+
+  const isLoadingPage =
+    blockLoadingStatus !== Loading.Complete && !blocks.length;
+  const isLoadingNext = blockLoadingStatus !== Loading.Complete;
+
+  // TODO: blocks are created every 32.768 seconds??
+  useEffect(() => {
+    const blockTimes = [];
+    let last: undefined | number;
+    for (const block of blocks) {
+      const blockCreationSeconds =
+        new Date(block.header.timestamp).getTime() / 1000;
+      if (last !== undefined) {
+        blockTimes.push(last - blockCreationSeconds);
+      }
+      last = blockCreationSeconds;
+    }
+
+    console.log({ blockTimes });
+  }, [blocks]);
+
+  useEffect(() => {
+    if (blockLoadingStatus === Loading.Idle) {
+      dispatch(fetchBlocks(blocksTableOptions));
+    }
+  }, []);
+
+  useEffect(() => {
+    if (refreshTimer === 0) {
+      dispatch(fetchBlocks(blocksTableOptions));
+    }
+  }, [refreshTimer]);
+
+  // TODO: make this more generic -> tie sorting to a store action maybe?
+  // possibly create filter/pagination/sorting (table config??) store?
+  useEffect(() => {
+    // TODO: should I add an if check here?
+    dispatch(fetchBlocks(blocksTableOptions));
+  }, [blocksTableOptions.sorting]);
+
   return (
-    <PageWrapper isLoading={isLoading}>
+    <PageWrapper isLoading={isLoadingPage}>
       <PageHead pageTitle={pageTitle} />
       <GradientHeading type="h2">{t('blocks')}</GradientHeading>
-      {blocks && (
-        <BlockTable
-          total={total}
-          blocks={blocks}
-          fetchMore={fetchNextPage}
-          isLoadingMoreBlocks={isFetchingNextPage}
-          sorting={sort}
-          onSortingChange={setSort}
-          initialSorting={initialSorting}
-        />
-      )}
+
+      <BlockTable
+        total={totalBlocks}
+        blocks={blocks}
+        fetchMore={() => {
+          dispatch(
+            fetchBlocks({
+              ...blocksTableOptions,
+              pagination: {
+                numToShow:
+                  blocksTableOptions.pagination.numToShow + DEFAULT_PAGINATION,
+              },
+            }),
+          );
+        }}
+        isLoadingMoreBlocks={isLoadingNext}
+        sorting={[
+          {
+            id: blocksTableOptions.sorting.sortBy,
+            desc: blocksTableOptions.sorting.order === 'desc',
+          },
+        ]}
+        onSortingChange={() =>
+          // TODO: will probably have a setOrdering/setSorting method that's less verbose
+          dispatch(
+            setPagination({
+              ...blocksTableOptions,
+              sorting: {
+                ...blocksTableOptions.sorting,
+                order:
+                  blocksTableOptions.sorting.order === 'desc' ? 'asc' : 'desc',
+              },
+            }),
+          )
+        }
+        initialSorting={initialSorting}
+      />
     </PageWrapper>
   );
 };
