+++ frontend/src/pages/index.ts
@@ -1,5 +1,6 @@
 export { AccountPage } from './Account';
 export { BlockPage } from './Block';
+export { Blocks as BlocksOld } from './BlocksOld';
 export { Blocks } from './Blocks';
 export { DeployPage } from './Deploy';
 export { Home } from './Home';
